Refactor to make usable outside of Rails
Before this change, some of the modules needed to use Dynamometer were
only available if the app/* directories were on the load path. This
isn't true by default outside of a Rails project.

Moves the actual modules from the concerns directories to the normal gem
lib directory. Also includes a method to coordinate attaching to
ActiveRecord::Base.